+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [Unreleased]
 ### Added
 - `rp.client.join.launch.timeout.value` and `rp.client.join.launch.timeout.unit` configuration properties to control SecondaryLaunch start timeout on client join, by @HardNorth
+### Changed
+- Disable Launch start wait for Secondary Launches if async reporting is enabled, by @HardNorth
+- Disable Statistics for Secondary Launches, by @HardNorth
 
 ## [5.2.15]
 ### Changed